+++ mp/src/game/client/hl2/hud_damageindicator.cpp
@@ -178,7 +178,7 @@ void CHudDamageIndicator::DrawDamageIndicator(int side)
 
 	int x1 = m_flDmgX;
 	int x2 = m_flDmgX + m_flDmgWide;
-	int y[4] = { m_flDmgY, m_flDmgY + insetY, m_flDmgY + m_flDmgTall1 - insetY, m_flDmgY + m_flDmgTall1 };
+	int y[4] = { (int)m_flDmgY, (int)(m_flDmgY + insetY), (int)(m_flDmgY + m_flDmgTall1 - insetY), (int)(m_flDmgY + m_flDmgTall1) };
 	int alpha[4] = { 0.0f, 1.0f, 1.0f, 0.0f };
 
 	// see if we're high damage
