+++ mp/src/game/shared/choreoevent.cpp
@@ -2076,8 +2076,8 @@ public:
 	{
 		if ( ARRAYSIZE( g_NameMap ) != CChoreoEvent::NUM_TYPES )
 		{
-			Error( "g_NameMap contains %i entries, CChoreoEvent::NUM_TYPES == %i!",
-				ARRAYSIZE( g_NameMap ), CChoreoEvent::NUM_TYPES );
+			Error( "g_NameMap contains %llu entries, CChoreoEvent::NUM_TYPES == %i!",
+				(uint64)(ARRAYSIZE( g_NameMap )), CChoreoEvent::NUM_TYPES );
 		}
 		for ( int i = 0; i < CChoreoEvent::NUM_TYPES; ++i )
 		{
@@ -2158,8 +2158,8 @@ public:
 	{
 		if ( ARRAYSIZE( g_CCNameMap ) != CChoreoEvent::NUM_CC_TYPES )
 		{
-			Error( "g_CCNameMap contains %i entries, CChoreoEvent::NUM_CC_TYPES == %i!",
-				ARRAYSIZE( g_CCNameMap ), CChoreoEvent::NUM_CC_TYPES );
+			Error( "g_CCNameMap contains %llu entries, CChoreoEvent::NUM_CC_TYPES == %i!",
+				(uint64)(ARRAYSIZE( g_CCNameMap )), CChoreoEvent::NUM_CC_TYPES );
 		}
 		for ( int i = 0; i < CChoreoEvent::NUM_CC_TYPES; ++i )
 		{
