+++ mp/src/materialsystem/stdshaders/spritecard.cpp
@@ -417,7 +417,7 @@ SHADER_DRAW
 		if ( bZoomSeq2 )
 		{
 			float flZScale=1.0/(params[ZOOMANIMATESEQ2]->GetFloatValue());
-			float C0[4]={ 0.5*(1.0+flZScale), flZScale, 0, 0 };
+			float C0[4]={ (float)(0.5*(1.0+flZScale)), flZScale, 0, 0 };
 			pShaderAPI->SetVertexShaderConstant( VERTEX_SHADER_SHADER_SPECIFIC_CONST_7, C0,
 				ARRAYSIZE(C0)/4 );
 		}
@@ -428,7 +428,7 @@ SHADER_DRAW
 
 		float VC0[8]={ params[MINSIZE]->GetFloatValue(), params[MAXSIZE]->GetFloatValue(),
 			params[STARTFADESIZE]->GetFloatValue(), params[ENDFADESIZE]->GetFloatValue(),
-			flStartFade, 1.0/(flMaxDistance-flStartFade),
+			flStartFade, (float)(1.0/(flMaxDistance-flStartFade)),
 			0,0 };
 
 		pShaderAPI->SetVertexShaderConstant( VERTEX_SHADER_SHADER_SPECIFIC_CONST_8, VC0, ARRAYSIZE(VC0)/4 );
